tableau_assistant/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
attrs==25.4.0
BCEmbedding==0.1.1
black==24.10.0
cachetools==6.2.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
chinesecalendar==1.11.0
click==8.3.1
colorama==0.4.6
coverage==7.12.0
cryptography==46.0.3
dataclasses-json==0.6.7
datasets==4.4.1
decorator==5.2.1
dill==0.4.0
distro==1.9.0
executing==2.2.1
faiss-cpu==1.13.0
fastapi==0.115.0
filelock==3.20.0
flake8==7.1.1
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hypothesis==6.148.3
idna==3.11
iniconfig==2.3.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json_repair==0.54.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-openai==1.1.0
langchain-anthropic==1.1.0
langchain-text-splitters==1.0.0
langchain_tableau==0.4.39
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langgraph-store==1.0.0
langsmith==0.4.48
deepagents==0.1.0
vizql-data-service-py==1.0.0
tableauserverclient==0.30
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mccabe==0.7.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy==1.13.0
mypy_extensions==1.1.0
networkx==3.6
numpy==2.3.5
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==22.0.0
pycodestyle==2.12.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.8.0
pyOpenSSL==25.3.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.38.6
sympy==1.14.0
tableauhyperapi==0.0.23576
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.32.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zhipuai==2.1.5.20250825
zstandard==0.25.0
